# docs/openapi-v2.yaml
openapi: 3.1.0
info:
  title: Vook Frontend API
  version: 2.0.0
  description: Resource-oriented contract consumed by the standalone Vook frontend.
servers: [{ url: /api/v2 }]
security: [{ sessionCookie: [] }]
paths:
...
  /attendance-devices:
    get: { operationId: listAttendanceDevices, responses: { '200': { $ref: '#/components/responses/Object' } } }
    post: { operationId: registerAttendanceDevice, requestBody: { $ref: '#/components/requestBodies/Object' }, responses: { '201': { $ref: '#/components/responses/Object' } } }
  /attendance-devices/{id}:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    put: { operationId: updateAttendanceDevice, requestBody: { $ref: '#/components/requestBodies/Object' }, responses: { '200': { $ref: '#/components/responses/Object' } } }
    delete: { operationId: deleteAttendanceDevice, responses: { '200': { $ref: '#/components/responses/Object' } } }
  /attendance-devices/{id}/events:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    get: { operationId: listAttendanceDeviceEvents, responses: { '200': { $ref: '#/components/responses/Object' } } }
...
components:
  securitySchemes:
    sessionCookie: { type: apiKey, in: cookie, name: vook_session }
  parameters:
    Id: { name: id, in: path, required: true, schema: { type: string } }
...
  requestBodies:
    Object: { required: true, content: { application/json: { schema: { type: object, additionalProperties: true } } } }
...
  responses:
...
    Object: { description: Success, content: { application/json: { schema: { $ref: '#/components/schemas/ObjectEnvelope' } } } }
...
  schemas:
...
    ListMeta:
      type: object
      required: [page, pageSize, total, totalPages]
      properties: { page: { type: integer, minimum: 1 }, pageSize: { type: integer, minimum: 1 }, total: { type: integer, minimum: 0 }, totalPages: { type: integer, minimum: 0 } }
...
    ObjectEnvelope: { type: object, required: [data], properties: { data: {}, meta: { $ref: '#/components/schemas/ListMeta' } } }
...